fix: extract shared worktree removal core
Keep dialog and MCP archive paths on the same removal flow so
retry handling, rollback, and notifications stay consistent.
